Save Clock Play/Pause state
Playing or Paused will reload as Paused (armed)

+++ software/src/HSClockManager.h
@@ -67,7 +67,7 @@ public:
     uint32_t clock_tick[2] = {0,0}; // previous ticks when a physical clock was received on DIGITAL 1
     uint32_t beat_tick = 0; // The tick to count from
     bool tock[NR_OF_CLOCKS] = {0,0,0,0,0,0,0,0,0}; // The current tock value
-    int16_t tocks_per_beat[NR_OF_CLOCKS] = {0,0, 0,0, 0,0, 0,0, MIDI_OUT_PPQN}; // Multiplier
+    int8_t tocks_per_beat[NR_OF_CLOCKS] = {0,0, 0,0, 0,0, 0,0, MIDI_OUT_PPQN}; // Multiplier
     int count[NR_OF_CLOCKS] = {0,0,0,0, 0,0,0,0, 0}; // Multiple counter, 0 is a special case when first starting the clock
     int8_t shuffle = 0; // 0 to 100
     int8_t shuffle_setting = 0;
@@ -118,7 +118,7 @@ public:
         tempo_setting = tempo = 1000000 / ticks_per_beat; // imprecise, for display purposes
     }
 
-    int GetMultiply(int ch = 0) {return tocks_per_beat[ch];}
+    int8_t GetMultiply(int ch = 0) {return tocks_per_beat[ch];}
     int GetClockPPQN() { return clock_ppqn; }
 
     void SetShuffle(int8_t sh_) { shuffle_setting = shuffle = constrain(sh_, 0, 99); }
